1. Changing totalCapacity & charge to floats instead of Ints
2. Adding current charge value to the UI

+++ api/prisma/migrations/20240607163001_create_battery_table/migration.sql
@@ -5,8 +5,8 @@ CREATE TYPE "BatteryStatus" AS ENUM ('Empty', 'Charged', 'Full');
 CREATE TABLE "Battery" (
     "id" SERIAL NOT NULL,
     "name" TEXT NOT NULL,
-    "totalCapacity" INTEGER NOT NULL DEFAULT 4,
-    "charge" INTEGER NOT NULL DEFAULT 4,
+    "totalCapacity" DOUBLE PRECISION NOT NULL DEFAULT 4,
+    "charge" DOUBLE PRECISION NOT NULL DEFAULT 4,
     "emptyCount" INTEGER NOT NULL DEFAULT 0,
     "status" "BatteryStatus" NOT NULL DEFAULT 'Full',
